+++ src/components/commons/loading/Loading.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './Loading.css';
+
+const Loading = () => (
+    <div className="lds-ring">
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+    </div>
+);
+
+export default Loading;
