+++ src/components/hacktoberfest/HacktoberfestPage.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Ranking from './Ranking';
+
+const HacktoberfestPage = () => {
+    const aa = '';
+    return (
+        <>
+            <h1>Hacktoberfest 2019</h1>
+            <Ranking />
+        </>
+    );
+};
+
+export default HacktoberfestPage;
